% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyconsole.R
\name{shinyconsole}
\alias{shinyconsole}
\title{Console messages from shiny}
\usage{
shinyconsole(session, rv = NULL)
}
\arguments{
\item{session}{A shiny session}

\item{rv}{A reactiveValues object}
}
\description{
Constantly observes the shiny session and reactive values (is provided), then
prints them to the console.
}
